Update files to make the GUI work like the JavaFX tutorial
