+++ src/main/java/knight2103/Launcher.java
@@ -1,3 +1,5 @@
+package knight2103;
+
 import javafx.application.Application;
 
 /**
