+++ src/main/java/knight2103/MainWindow.java
@@ -1,3 +1,5 @@
+package knight2103;
+
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
@@ -18,10 +20,12 @@ public class MainWindow extends AnchorPane {
     @FXML
     private Button sendButton;
 
-    private Duke duke;
+    private Knight2103 knight2103;
 
-    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
-    private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
+    private Image userImage = new Image(
+            this.getClass().getResourceAsStream("/images/Koobit_Jaguar.png"));
+    private Image dukeImage = new Image(
+            this.getClass().getResourceAsStream("/images/Koobit_Water_Droplet.png"));
 
     @FXML
     public void initialize() {
@@ -29,8 +33,8 @@ public class MainWindow extends AnchorPane {
     }
 
     /** Injects the Duke instance */
-    public void setDuke(Duke d) {
-        duke = d;
+    public void setKnight2103(Knight2103 chatbot) {
+        knight2103 = chatbot;
     }
 
     /**
@@ -40,7 +44,7 @@ public class MainWindow extends AnchorPane {
     @FXML
     private void handleUserInput() {
         String input = userInput.getText();
-        String response = duke.getResponse(input);
+        String response = knight2103.getResponse(input);
         dialogContainer.getChildren().addAll(
                 DialogBox.getUserDialog(input, userImage),
                 DialogBox.getDukeDialog(response, dukeImage)
