+++ src/main/resources/view/DialogBox.fxml
@@ -1,77 +1,23 @@
-import javafx.fxml.FXML;
-        import javafx.geometry.Pos;
-        import javafx.scene.control.Label;
-        import javafx.scene.image.Image;
-        import javafx.scene.image.ImageView;
-        import javafx.scene.layout.HBox;
+<?xml version="1.0" encoding="UTF-8"?>
 
-        // Level 3c: allows "flipping" - align left
-        import javafx.collections.FXCollections;
-        import javafx.collections.ObservableList;
-        import javafx.scene.Node;
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.HBox?>
 
-        import java.io.IOException;
-        import java.util.Collections;
-        import javafx.fxml.FXMLLoader;
-
-        public class DialogBox extends HBox {
-        @FXML
-        private Label dialog;
-        @FXML
-        private ImageView displayPicture;
-
-        private void flip() {
-
-        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
-    Collections.reverse(tmp);
-    getChildren().setAll(tmp);
-    setAlignment(Pos.TOP_LEFT);
-
-    /*
-    this.setAlignment(Pos.TOP_LEFT);
-    ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
-        FXCollections.reverse(tmp);
-        this.getChildren().setAll(tmp);
-        */
-        }
-
-        // no change
-        public static DialogBox getUserDialog(String text, Image img) {
-        return new DialogBox(text, img);
-        }
-
-        // no change
-        public static DialogBox getDukeDialog(String text, Image img) {
-        var db = new DialogBox(text, img);
-        db.flip();
-        return db;
-        }
-
-        private DialogBox(String text, Image img) {
-        try {
-        FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
-        fxmlLoader.setController(this);
-        fxmlLoader.setRoot(this);
-        fxmlLoader.load();
-        } catch (IOException e) {
-        e.printStackTrace();
-        }
-
-        dialog.setText(text);
-        displayPicture.setImage(img);
-
-        /*
-        dialog = new Label(s);
-        displayPicture = new ImageView(i);
-
-        //Styling the dialog box
-        dialog.setWrapText(true);
-        displayPicture.setFitWidth(100.0);
-        displayPicture.setFitHeight(100.0);
-        this.setAlignment(Pos.TOP_RIGHT);
-
-        this.getChildren().addAll(dialog, displayPicture);
-        */
-        }
-        }
+<fx:root alignment="TOP_RIGHT"
+         maxHeight="1.7976931348623157E308"
+         maxWidth="1.7976931348623157E308"
+         prefWidth="400.0"
+         type="javafx.scene.layout.HBox"
+         xmlns="http://javafx.com/javafx/17"
+         xmlns:fx="http://javafx.com/fxml/1">
+    <children>
+        <Label fx:id="dialog" text="Label" wrapText="true" minHeight="-Infinity"/> <!-- important for vertical size expansion -->
+        <ImageView fx:id="displayPicture" fitHeight="99.0" fitWidth="99.0" pickOnBounds="true" preserveRatio="true" />
+    </children>
+    <padding>
+        <Insets bottom="15.0" left="5.0" right="5.0" top="15.0" />
+    </padding>
+</fx:root>
